Revised protocol to be derivative of JKKX16 with secret sharing over Curve25519

+++ android/src/main/java/xyz/juicebox/sdk/internal/Native.java
@@ -26,8 +26,8 @@ public final class Native {
 
     public static native long configurationCreate(
             @NotNull Realm[] realms,
-            byte registerThreshold,
-            byte recoverThreshold,
+            int registerThreshold,
+            int recoverThreshold,
             @NotNull PinHashingMode pinHashingMode);
 
     public static native boolean configurationsAreEqual(long configuration1, long configuration2);
